First round of architecture coding done
Not tested (at all), but, the basic architecture is coded up.

There are several major components:

1) Coordinator node. This node coordinates everything by partitioning
the graph and distributing it among worker nodes, and also serves as
a barrier between steps.

2) Worker node. This node manages the running of vertex programs for the
set of vertices assigned to it.

The worker also has several services:
A) The worker service, which is what the coordinator interacts with
B) The postoffice service, which handles delivery of vertex program
messages
C) Vertex actors. Not entirely thought through yet, but, the idea
is to have vertex programs run as actors to take advantage of multithreading
I still need to figure out what that means as far as exposing various elements
of the system to remote access.