Fixed a race condition, passing specs now
There was a race condition on reading mail boxes:
One node could end up sending messages before a previous node has
read the destinations mailbox.

This is being dealt with via a separate "init_superstep" method that
reads all mailboxes prior to the "superstep" method being invoked on
workers.

This is probably not the ideal solution...