Fixed (i think) marshall dump errors
I believe that these were failing because i was inadvertantly returning
a vertex's mailbox when i did a remote push.

The problem was (I think) that a fiber could be suspended between
adding the message to the mailbox and returning the result.

Thus, the mailbox could be modified prior to the return finishing.

I.e., it could be modified before the marshall finishes.

I'm not sure if this is true, but I can't get it to fail with a 2 node
local setup anymore.