Triad census fix attempt.
There is a local computation for the type 1s.

This is because a vertex would need to know the total number of
vertices in the graph to compute the type 1s it's a member of.

Thus, the type 1s have to be determined the same was as the
type 0s