Performance mega patch
Post office message buffering:
Post office out boxes that are buffered with messages sent in bulk
when they exceed the buffer size. This results in a pretty significant
performance increase as it drastically reduces the number of mesasges
sent between compute nodes.

Worker graph partitions:
Currently runs the vertex programs in two seperate threads. Old
method put each vertex in its own thread (or what amounts to a
thread in dcell). This will probably need to be worked on more in the
future.

Graph distribution in threads:
Currently tries to distribute the graph by splitting it up into
1k node slices that are distributed in parallel. Each vertex
is still added to the worker individually; not in bulk.