Improved performance (hopefully) of Worker#active?
Old version was looping through (possibly) all vertices and
mailboxes to see if the worker had any work left.

New version deals with this "dynamically" by adding an @active
member variable (set to true on initialization).

in #init_superstep, the worker will set its active state to false.
If there are any active vertices (i.e., thsoe that have not yet
voted to stop or have messages waiting), #active! will be called.

Additionally, if a message is received during a step, the postoffice will set the
worker to active by calling Worker#active!

This should ensure that after the superstep barrier is reached,
all workers that have work to do at step t + 1 will be active,
while the rest will not be active.